Add repository documentation data source
With this commit it becomes possible to define additional data sources
for content pages. The content will be cloned from the specified
repository. Links in these pages will be normalized to work inside the
docs project. It's assumed the documentation pages include the standard
nanoc config options and set :title and optionall :sort_rank.